+++ src/app/services/supabaseChatbot.ts
@@ -26,6 +26,7 @@ export const reviewApi = {
       .order("created_at", { ascending: false });
 
     if (error) throw error;
+
     return data as ReviewType[];
   }
 };
